+++ script/item/items.ts
@@ -538,6 +538,48 @@ const armors = {
     type: "armor",
     slot: "legs",
   },
+  /* Orc Chief set */
+  orc_chief_headdress: {
+    id: "orc_chief_headdress",
+    price: 1080,
+    speed: 0,
+    defence: {
+      physical: 8,
+      magical: 5,
+      elemental: 4,
+    },
+    modifiers: {
+      strV: 3,
+      vitV: 3,
+      agiV: 3,
+      hpMaxP: 5,
+    },
+    stackable: true,
+    icon: "gfx/items/horned-helm.png",
+    tier: itemTiers.epic,
+    type: "armor",
+    slot: "helmet",
+  },
+  orc_chief_armor: {
+    id: "orc_chief_armor",
+    price: 2000,
+    speed: 0,
+    defence: {
+      physical: 18,
+      magical: 10,
+      elemental: 10,
+    },
+    modifiers: {
+      hpMaxV: 20,
+      hpMaxP: 7,
+      strP: 5,
+    },
+    stackable: true,
+    icon: "gfx/items/leather-vest.png",
+    tier: itemTiers.epic,
+    type: "armor",
+    slot: "armor",
+  },
   /* Acolyte Set */
   acolyte_hood: {
     id: "acolyte_hood",